# docs/master_swagger.yaml
basePath: /api/
definitions:
  api.GroupResult:
    properties:
      id:
        type: string
      name:
        type: string
      total:
        type: number
    type: object
...
  curd.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  curd.ReplyData-api_GroupResult:
    properties:
      data:
        $ref: '#/definitions/api.GroupResult'
      error:
        type: string
    type: object
...
  curd.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  curd.ReplyData-model_Aggregator:
    properties:
      data:
        $ref: '#/definitions/types.Aggregator'
      error:
        type: string
    type: object
  curd.ReplyData-model_Alarm:
    properties:
      data:
        $ref: '#/definitions/types.Alarm'
      error:
        type: string
    type: object
...
  curd.ReplyList-model_Aggregator:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Aggregator'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-model_Alarm:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Alarm'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.Aggregator:
    properties:
      assign:
        description: 赋值
        type: string
      created:
        type: string
      crontab:
        description: 定时计划
        type: string
      desc:
        description: 说明
        type: string
      device_id:
        type: string
      disabled:
        description: 禁用
        type: boolean
      expression:
        description: 表达式
        type: string
      id:
        type: string
      name:
        description: 名称
        type: string
      product_id:
        type: string
      type:
        description: 聚合算法 inc dec avg count min max sum last first
        type: string
    type: object
  types.Alarm:
    properties:
      created:
        type: string
      device:
        type: string
      device_id:
        type: string
      id:
        type: integer
      level:
        type: integer
      message:
        type: string
      product:
        type: string
      product_id:
        type: string
      read:
        type: boolean
      title:
        type: string
      type:
        type: string
      validator_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师接口文档
  version: 3.2 版本
paths:
  /aggregator/{id}:
    get:
      consumes:
      - application/json
      description: 获取聚合器
      parameters:
      - description: 聚合器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Aggregator'
      summary: 获取聚合器
      tags:
      - aggregator
    post:
      consumes:
      - application/json
      description: 修改聚合器
      parameters:
      - description: 聚合器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 聚合器信息
        in: body
        name: aggregator
        required: true
        schema:
          $ref: '#/definitions/types.Aggregator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Aggregator'
      summary: 修改聚合器
      tags:
      - aggregator
  /aggregator/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除聚合器
      parameters:
      - description: 聚合器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Aggregator'
      summary: 删除聚合器
      tags:
      - aggregator
  /aggregator/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用聚合器
      parameters:
      - description: 聚合器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Aggregator'
      summary: 禁用聚合器
      tags:
      - aggregator
  /aggregator/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用聚合器
      parameters:
      - description: 聚合器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Aggregator'
      summary: 启用聚合器
      tags:
      - aggregator
  /aggregator/count:
    post:
      consumes:
      - application/json
      description: 查询聚合器数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询聚合器数量
      tags:
      - aggregator
  /aggregator/create:
    post:
      consumes:
      - application/json
      description: 创建聚合器
      parameters:
      - description: 聚合器信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Aggregator'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Aggregator'
      summary: 创建聚合器
      tags:
      - aggregator
...
  /aggregator/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入聚合器
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入聚合器
      tags:
      - aggregator
  /aggregator/list:
    get:
      consumes:
      - application/json
      description: 查询聚合器
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Aggregator'
      summary: 查询聚合器
      tags:
      - aggregator
  /aggregator/search:
    post:
      consumes:
      - application/json
      description: 查询聚合器
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Aggregator'
      summary: 查询聚合器
      tags:
      - aggregator
  /alarm/{id}/delete:
    get:
      description: 删除报警
      parameters:
      - description: 报警ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Alarm'
      summary: 删除报警
      tags:
      - alarm
  /alarm/{id}/read:
    get:
      description: 阅读报警
      parameters:
      - description: 报警ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Alarm'
      summary: 阅读报警
      tags:
      - alarm
  /alarm/count:
    post:
      consumes:
      - application/json
      description: 查询报警
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询报警
      tags:
      - alarm
  /alarm/list:
    get:
      description: 查询报警
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Alarm'
      summary: 查询报警
      tags:
      - alarm
  /alarm/search:
    post:
      consumes:
      - application/json
      description: 查询报警
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Alarm'
      summary: 查询报警
      tags:
      - alarm
...
  /history/group:
    get:
      description: 原始数据按时间统计
      parameters:
      - description: 设备区域
        in: query
        name: area
        type: string
      - description: 结束时间
        in: query
        name: end
        type: string
      - description: 设备分组
        in: query
        name: group
        type: string
      - description: 数据点位
        in: query
        name: point
        type: string
      - description: 起始时间
        in: query
        name: start
        type: string
      - description: 设备类型
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-api_GroupResult'
      summary: 原始数据按时间统计
      tags:
      - group
...
swagger: "2.0"